Backend/api-IAFraturas/requirements.txt
fastapi
uvicorn
tensorflow==2.17.0
numpy<2.0.0
pillow
pydicom
firebase-admin
python-multipart
requests
